Add files via upload
What is WebPhisher? Webphisher is a PyPhisher made for Windows and Mac OS without using Termux. What does our script do? It creates phishing links to the social network of your choice. Instagram Facebook, Github, Gmail, Instagram, Netflix, Paypal, PlayStation, Reddit, Spotify, Steam, Tiktok, Twitch, Twitter, VK, Yandex. You have 17 social networks to choose from. 

How do I work with the script? 
To start, download the zip archive 'Webphisher-main' and open 'Webphisher.py '. 
Next, enter the number that is presented in front of the social network that you want to open. In my case, it's Yandex. We are waiting for the link to be created. We have a link ready, but don't rush to follow it! First, open the 'Social Networks' folder, and look for the folder with the social network that you have selected. Open the file .html in the IDE. And then click on the link. Ready!